requirements.txt
torch
pytorch-lightning==1.9.4
sconf
Pillow
timm
transformers
zss
sentencepiece
nltk
datasets[vision]
tensorboard
wandb
typed-argument-parser
black
gradio
